EXAMPLES: make the lru test code work without the optional arg
When lru is built in standalone mode for testing, starting it indicates
that the second argument is optional, but when not passed, the lru head
is not initialized and the first get crashes. Let's assign a default
value in this case. We can also shut a warning that triggers in case
it's not properly initialized.