TESTS: bench: also count dels and tree elems
Performance without any idea of the number of elements is meaningless.
Let's count dels and real inserts to figure how many entries are in the
tree. We now report lookups, ins, del, and only per second since it's
pointless to report the instant count based on an unknown time value.
Lines are slightly shorter and more readable. Note that failed inserts
due to existing values failed to be inserted into a unique tree are
reported as lookups (since it's basically what they really are).