TESTS: finally drop the bench utility
It was not a good idea to integrate the tool from cebtree, it takes
more development time to keep it in sync between the two sub-projects
than to focus on the code. Let's drop it and keep it in a separate
project, which will concentrate all the benchmarking tools instead.

This means that the tests/ directory is now empty. Let's just keep
the implicit rule allowing to build anything quickly there from a
dirty repository and nothing else.