OPTIM: ebst: optimize insert for modern CPUs
Similar to previous patches, let's improve the insert() descent loop to
avoid discovering mandatory data too late. The key is checked against a
mask that is calculated as soon as possible, and troot is calculated a
bit before the end of the loop. We try to no longer use the side as an
array but as a ternary evaluation instead so that the addresses of the
two branches can be speculated early by the CPU.

All this results in 41..51% performance improvement on small and medium
strings.